Small adjustments for Clippy

+++ ceno_zkvm/src/instructions/riscv/mulh.rs
@@ -532,7 +532,7 @@ mod test {
             .unwrap();
 
         let signed_unsigned_prod_high =
-            (((rs1 as i32) as i64).wrapping_mul(rs2 as i64) >> 32) as u32;
+            ((rs1 as i64).wrapping_mul(rs2 as i64) >> 32) as u32;
 
         // values assignment
         let insn_code = encode_rv32(InsnKind::MULHSU, 2, 3, 4, 0);
@@ -543,7 +543,7 @@ mod test {
                     MOCK_PC_START,
                     insn_code,
                     rs1 as u32,
-                    rs2 as u32,
+                    rs2,
                     Change::new(0, signed_unsigned_prod_high),
                     0,
                 ),
